Refactor: `sms`에서 `email`로 파일명 변경
