+++ src/api/index.ts
@@ -2,7 +2,7 @@ export * from './login';
 export * from './application';
 export * from './applicationForm';
 export * from './team';
-export * from './sms';
+export * from './email';
 export * from './generation';
 export * from './member';
 export * from './scoreHistory';
