+++ src/components/modal/EmailSendDetailInfoModalDialog/EmailSendDetailInfoModalDialog.styled.ts
