+++ src/components/modal/index.ts
@@ -1,5 +1,5 @@
 export { default as ChangeResultModalDialog } from './ChangeResultModalDialog/ChangeResultModalDialog.component';
-export { default as EmailSendDetailInfoModalDialog } from './SmsSendDetailInfoModalDialog/SmsSendDetailInfoModalDialog.component';
-export { default as EmailSendDetailListModalDialog } from './SmsSendDetailListModalDialog/SmsSendDetailListModalDialog.component';
+export { default as EmailSendDetailInfoModalDialog } from './EmailSendDetailInfoModalDialog/EmailSendDetailInfoModalDialog.component';
+export { default as EmailSendDetailListModalDialog } from './EmailSendDetailListModalDialog/EmailSendDetailListModalDialog.component';
 export { default as ApplyActivityScoreModalDialog } from './ApplyActivityScoreModalDialog/ApplyActivityScoreModalDialog.component';
 export { default as ActivityScoreModalDialog } from './ActivityScoreModalDialog/ActivityScoreModalDialog.component';
