+++ src/pages/EmailSendingList/EmailSendingList.styled.ts
