+++ src/store/index.ts
@@ -4,5 +4,5 @@ export * from './application';
 export * from './team';
 export * from './applicationForm';
 export * from './toast';
-export * from './sms';
+export * from './email';
 export * from './generation';
