+++ src/types/dto/email.ts
