+++ src/types/dto/index.ts
@@ -3,7 +3,7 @@ export * from './base';
 export * from './application';
 export * from './applicationForm';
 export * from './team';
-export * from './sms';
+export * from './email';
 export * from './generation';
 export * from './member';
 export * from './scoreHistory';
